requirements.txt
flask==3.0.3
flask-cors==4.0.0
google-analytics-data==0.18.3
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
